backend/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
python-dotenv
fastmcp
bcrypt
python-jose
